requirements.txt
alembic==1.6.5
click==8.0.1
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
lazy-object-proxy==1.4.0
mccabe==0.6.1
pycryptodome==3.7.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.12.0
typed-ast==1.4.2
psycopg2-binary==2.9.9
SQLAlchemy==1.4.18
Werkzeug==2.0.1
python-dotenv==0.11.0